% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{add.dependency}
\alias{add.dependency}
\title{Private function that adds a package to the list of dependencies
that will be installed on the Promote server}
\usage{
add.dependency(name, importName, src, version, install, auth_token, branch)
}
\arguments{
\item{name}{name of the package to be installed}

\item{importName}{name under which the package is imported (for a github package,
this may be different from the name used to install it)}

\item{src}{source that the package is installed from (CRAN or github)}

\item{version}{version of the package}

\item{install}{whether or not the package should be installed in the model image}

\item{auth_token}{a personal access token for github or gitlab repositories}

\item{branch}{The branch or tag of the package to be installed}
}
\description{
Private function that adds a package to the list of dependencies
that will be installed on the Promote server
}
