% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add.metadata}
\alias{add.metadata}
\title{Private function that adds metadata about the model
that will be installed on the Promote server
the metadata is arranged as key-value pairs}
\usage{
add.metadata(key, value)
}
\arguments{
\item{key}{key name for the metadata entry}

\item{value}{value for the metadata entry}
}
\description{
Private function that adds metadata about the model
that will be installed on the Promote server
the metadata is arranged as key-value pairs
}
