% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-capture.R
\name{capture.src}
\alias{capture.src}
\title{Private function for catpuring the source code of model}
\usage{
capture.src(funcs, capture.model.require = TRUE)
}
\arguments{
\item{funcs}{functions to capture, defaults to required promote model functions}

\item{capture.model.require}{flag to capture the model.require function}
}
\description{
Private function for catpuring the source code of model
}
