% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-utils.R
\name{promote.get}
\alias{promote.get}
\title{Private function for performing a GET request}
\usage{
promote.get(endpoint, query = c())
}
\arguments{
\item{endpoint}{/path for REST request}

\item{query}{url parameters for request}
}
\description{
Private function for performing a GET request
}
