% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-capture.R
\name{promote.ls}
\alias{promote.ls}
\title{Private function for determining model dependencies}
\usage{
promote.ls()
}
\description{
List all object names which are dependencies of and `model.predict`.
}
