% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote.R
\name{promote.metadata}
\alias{promote.metadata}
\title{Add metadata to the deployment of your promote model}
\usage{
promote.metadata(name, value)
}
\arguments{
\item{name}{key name for the metadata entry}

\item{value}{value for the metadata entry}
}
\description{
Add metadata to the deployment of your promote model
}
\examples{
\dontrun{
promote.metadata("key", "value")
promote.metadata("R_squared", summary(fit)$r.squared)
promote.metadata("R_squared_adj", summary(fit)$adj.r.squared)
promote.metadata("deploy_node", Sys.info()[["nodename"]])
}
}
\keyword{metadata}
