% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-utils.R
\name{promote.post}
\alias{promote.post}
\title{Private function for performing a POST request}
\usage{
promote.post(endpoint, query = c(), data, silent = TRUE, bulk = FALSE)
}
\arguments{
\item{endpoint}{/path for REST request}

\item{query}{url parameters for request}

\item{data}{payload to be converted to raw JSON}

\item{silent}{should output of url to console be silenced?
Default is \code{FALSE}.}

\item{bulk}{is this a bulk style request? Default is \code{FALSE}.}
}
\description{
Private function for performing a POST request
}
