% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote.R
\name{promote.predict}
\alias{promote.predict}
\title{Make a prediction using promote.}
\usage{
promote.predict(model_name, data, model_owner, raw_input = FALSE,
  silent = TRUE)
}
\arguments{
\item{model_name}{the name of the model you want to call}

\item{data}{input data for the model}

\item{model_owner}{the owner of the model [optional]}

\item{raw_input}{when true, incoming data will NOT be coerced into data.frame}

\item{silent}{should output of url to console (via \code{promote.post})
be silenced? Default is \code{FALSE}.}
}
\description{
This function calls promote's REST API and returns a response formatted as a
data frame.
}
\examples{
promote.config <- c(
 username = "your username",
 apikey = "your apikey",
 env = "http://sandbox.promotehq.com/"
)
\dontrun{
promote.predict("irisModel", iris)
}
}
\keyword{predict}
