% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-capture.R
\name{promote.spider.block}
\alias{promote.spider.block}
\title{Private function for recursively looking for variables}
\usage{
promote.spider.block(block, defined.vars = c())
}
\arguments{
\item{block}{code block to spider}

\item{defined.vars}{variables which have already been defined within the
scope of the block. e.g. function argument}
}
\description{
Private function for recursively looking for variables
}
