% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-capture.R
\name{promote.spider.func}
\alias{promote.spider.func}
\title{Private function for spidering function source code}
\usage{
promote.spider.func(func.name)
}
\arguments{
\item{func.name}{name of function you want to spider}
}
\description{
Private function for spidering function source code
}
