% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promote.R
\name{promote.unload}
\alias{promote.unload}
\title{Removes a library from the promote model's dependency list}
\usage{
promote.unload(name)
}
\arguments{
\item{name}{of the package to be removed}
}
\description{
Removes a library from the promote model's dependency list
}
\examples{
\dontrun{
promote.unload("wesanderson")
}
}
